Allow disabling of `rawError` by default
Sometimes the wrapped Action errors can be noisy (for example, if the
error has nothing to do with `vuex-module-decorators`, which is more
often than not the case).

This is a non-breaking change that allows `rawError` to be set to `true`
by default in all `@Action` decorators.

We achieve this by adding a global, module-wide `config` object, which
can have values set like so:

```typescript
import { config } from 'vuex-module-decorators
// Set rawError to true by default on all @Action decorators
config.rawError = true
```