enable inheritance and extension of the state object
Once a module was defined with a state object, extending it did not
change that state object.  Both fields and defaults were affected.

Explicitly check if a state function is defined, to avoid overwriting
a function with an object, but overwrite the state object with the
extended object if it is not.