✨ Allow getters to be reconfigured
At the moment, getters cannot be stubbed in tests, because they're
defined as unconfigurable.

This change adds a `configurableGetters` option to the `Module` options,
which is `false` by default (keeps existing behaviour).

If this flag is set to `true`, we allow getters to be reconfigured,
which lets us stub them.